more superficial refactoring
