+++ userspace/libsinsp/fields_info.cpp
@@ -28,7 +28,6 @@ limitations under the License.
 
 #include <sinsp.h>
 #include "fields_info.h"
-#include "source_plugin.h"
 
 // Must match the value in the zsh tab completion
 #define DESCRIPTION_TEXT_START 16
