+++ userspace/libsinsp/plugin.h
